EcoNexo/mcp-server/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
mcp>=0.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0

duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
